odd and even count

+++ intro_prog_lang_prob_solv_part_1/weired_algorithm.c
@@ -2,22 +2,22 @@
 #include <stdio.h>
 int main()
 {
-    int n;
-    scanf("%d", &n);
-    printf("%d ", n);
+    long long int n;
+    scanf("%lld", &n);
+    printf("%lld ", n);
     while (n != 1)
     {
         if (n % 2 == 0)
         {
             n = n / 2;
-            printf("%d ", n);
+            printf("%lld ", n);
         }
         else
         {
             if (n % 2 != 0)
             {
                 n = (n * 3) + 1;
-                printf("%d ", n);
+                printf("%lld ", n);
             }
         }
     }
